Update sana_SW02-03
This adds the template for the SW02-01 which is the single switch version of this switch.  There is also a two gang version that i am sure just removes one of the other relays but as i don't have it i can't confirm which one it is.  This makes it cleaner than having to use the 03 template and editing to remove